feat: add name to auth0 scope to support storing name to the user in db
