+++ packages/client/app/src/utils/AuthService.js
@@ -15,7 +15,7 @@ export default class AuthService {
                 redirectUrl: config.api.redirectURL,
                 responseType: "token",
                 params: {
-                    scope: "openid roles user_id email"
+                    scope: "openid roles user_id email profile"
                 }
             },
             theme: {
